Move all python source files to /src/effortsharing
