+++ src/effortsharing/class_allocation.py
