+++ tests/profiling.py
@@ -7,7 +7,7 @@ import os
 parent_dir = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
 sys.path.insert(0, parent_dir)
 
-from class_allocation import allocation
+from effortsharing.class_allocation import allocation
 
 def profile_allocation():
     '''
